+++ PreOrclFrontEnd/Models/Usuarios.cs
@@ -11,7 +11,7 @@ namespace PreOrclFrontEnd.Models
     public class Usuarios
     {
         [Key]
-        public int IdUsuario { get; set; }
+        public decimal IdUsuario { get; set; }
 
         [Required(ErrorMessage = "Campo Requerido")]
         public string Usuario { get; set; }
